Criando os cruds criar, lista, excruir clientes

+++ app_clientes_console/src/models/cliente.rs
@@ -1,6 +1,7 @@
-#[allow(dead_code)]
+#[derive(Debug, Default)]
+
 pub struct Cliente {
-    pub id: u32,
+    pub id: usize,
     pub nome: String,
     pub cpf: String,
     pub endereco: String,
